chore: clean up some post-merge issues
